Update docker-compose dev flow
By mounting the current directoy inside the app container we can develop
as though everything was native. The improved VS Code tasks make use of
the docker-compose run command to run the server inside docker, but
expose it to the host's port 4000.

This might mess things up for production deploys. Mounting the current
dir as a volume may not work for production. Instead, we may want to
pull this out into the task commands. But :shrug:.